samples/s3-lambda-trigger/resources/lambda/fn_pgp_decrypt/requirements.txt
slalom.dataops
boto3
gnupg
